feat(services): add check user role for crud post

+++ src/controllers/post.controller.js
@@ -5,12 +5,38 @@ import {
   updateOneWithPatch,
   deleteOne,
   findAllCommentsByPostId,
+  checkSecurityAccessRessource,
 } from "../services/post.service.js";
 
 import AppError from "../utils/appError.js";
 
+export const getAllPostsAsAdmin = async (req, res, next) => {
+  try {
+    const {
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
+
+    await checkSecurityAccessRessource("readAllAsAdmin", currentUserId);
+
+    const posts = await findAll(req.query, true);
+    res.status(200).json(posts);
+  } catch (error) {
+    next(error);
+  }
+};
+
 export const getAllPosts = async (req, res, next) => {
   try {
+    const {
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
+
+    await checkSecurityAccessRessource("readAll", currentUserId);
+
     const posts = await findAll(req.query);
     res.status(200).json(posts);
   } catch (error) {
@@ -19,19 +45,25 @@ export const getAllPosts = async (req, res, next) => {
 };
 
 export const createPost = async (req, res, next) => {
-  const { title, content, isPublished, createdAt, updatedAt, authorId } =
-    req.body;
-
-  const datas = {
-    title,
-    content,
-    isPublished,
-    createdAt: createdAt || new Date(),
-    updatedAt: updatedAt || new Date(),
-    authorId,
-  };
-
   try {
+    const {
+      body: { title, content, isPublished, createdAt, updatedAt, authorId },
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
+
+    await checkSecurityAccessRessource("create", currentUserId);
+
+    const datas = {
+      title,
+      content,
+      isPublished,
+      createdAt: createdAt || new Date(),
+      updatedAt: updatedAt || new Date(),
+      authorId,
+    };
+
     const post = await createOne(datas);
     res.status(201).json(post);
   } catch (error) {
@@ -41,7 +73,14 @@ export const createPost = async (req, res, next) => {
 
 export const getPost = async (req, res, next) => {
   try {
-    const postId = req.params.id;
+    const {
+      params: { id: postId },
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
+
+    await checkSecurityAccessRessource("readOne", currentUserId, postId);
 
     if (!postId || !Number(postId)) {
       throw new AppError(404, "fail", "Missing post id");
@@ -56,18 +95,24 @@ export const getPost = async (req, res, next) => {
 };
 
 export const updatedPost = async (req, res, next) => {
-  const { title, content, isPublished, updatedAt, authorId } = req.body;
-
-  const datas = {
-    title,
-    content,
-    isPublished,
-    updatedAt: updatedAt || new Date(),
-    authorId,
-  };
-
   try {
-    const postId = req.params.id;
+    const {
+      body: { title, content, isPublished, updatedAt, authorId },
+      params: { id: postId },
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
+
+    await checkSecurityAccessRessource("update", currentUserId, postId);
+
+    const datas = {
+      title,
+      content,
+      isPublished,
+      updatedAt: updatedAt || new Date(),
+      authorId,
+    };
 
     if (!postId || !Number(postId)) {
       throw new AppError(404, "fail", "Missing post id");
@@ -83,11 +128,18 @@ export const updatedPost = async (req, res, next) => {
 
 export const deletePost = async (req, res, next) => {
   try {
-    const postId = req.params.id;
+    const {
+      params: { id: postId },
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
 
     if (!postId || !Number(postId)) {
       throw new AppError(404, "fail", "Missing post id");
     }
+
+    await checkSecurityAccessRessource("delete", currentUserId, postId);
 
     await deleteOne(postId);
 
@@ -103,7 +155,14 @@ export const deletePost = async (req, res, next) => {
 
 export const getAllPostComments = async (req, res, next) => {
   try {
-    const postId = req.params.id;
+    const {
+      params: { id: postId },
+      session: {
+        user: { id: currentUserId },
+      },
+    } = req;
+
+    await checkSecurityAccessRessource("readOne", currentUserId, postId);
 
     if (!postId || !Number(postId)) {
       throw new AppError(404, "fail", "Missing post id");
